+++ src/PkgREPL.jl
@@ -26,7 +26,7 @@ $
     CondaPkg.PkgSpec(name, version = version, channel = channel, build = build)
 end
 
-function parse_pip_pkg(x::String; binary::String = "")
+function parse_pip_pkg(x::String; binary::String = "", editable=false)
     m = match(
         r"""
 ^
@@ -41,7 +41,7 @@ $
     name = m.captures[1]
     extras = split(something(m.captures[3], ""), ",", keepempty = false)
     version = something(m.captures[4], "")
-    CondaPkg.PipPkgSpec(name, version = version, binary = binary, extras = extras)
+    CondaPkg.PipPkgSpec(name, version = version, binary = binary, extras = extras, editable = editable)
 end
 
 function parse_channel(x::String)
@@ -60,6 +60,13 @@ const binary_opt = Pkg.REPLMode.OptionDeclaration([
     :name => "binary",
     :takes_arg => true,
     :api => :binary => identity,
+])
+
+const editable_opt = Pkg.REPLMode.OptionDeclaration([
+    :name => "editable",
+    :short_name => "e",
+    :takes_arg => false,
+    :api => :editable => true,
 ])
 
 ### status
@@ -191,13 +198,13 @@ const channel_add_spec = Pkg.REPLMode.CommandSpec(
 
 ### pip_add
 
-function pip_add(args; binary = "")
-    CondaPkg.add([parse_pip_pkg(arg, binary = binary) for arg in args])
+function pip_add(args; binary = "", editable = false)
+    CondaPkg.add([parse_pip_pkg(arg, binary = binary, editable = editable) for arg in args])
 end
 
 const pip_add_help = Markdown.parse("""
 ```
-conda pip_add [--binary={only|no}] pkg ...
+conda pip_add [--binary={only|no}] [--editable] pkg ...
 ```
 
 Add Pip packages to the environment.
@@ -209,6 +216,7 @@ pkg> conda pip_add build
 pkg> conda pip_add build~=0.7                # version range
 pkg> conda pip_add pydantic[email,timezone]  # extras
 pkg> conda pip_add --binary=no nmslib        # always build from source
+pkg> conda pip_add --editable nmslib@/path/to/package_source  # install locally
 ```
 """)
 
@@ -219,7 +227,7 @@ const pip_add_spec = Pkg.REPLMode.CommandSpec(
     help = pip_add_help,
     description = "add Pip packages",
     arg_count = 0 => Inf,
-    option_spec = [binary_opt],
+    option_spec = [binary_opt, editable_opt],
 )
 
 ### rm