+++ src/resolve.jl
@@ -518,6 +518,7 @@ function _resolve_merge_pip_packages(packages)
         urls = String[]
         binary = ""
         extras = String[]
+        editable = false
         for (fn, pkg) in pkgs
             @assert pkg.name == name
             if startswith(pkg.version, "@")
@@ -539,6 +540,7 @@ function _resolve_merge_pip_packages(packages)
                 end
             end
             append!(extras, pkg.extras)
+            editable |= pkg.editable
         end
         sort!(unique!(urls))
         sort!(unique!(versions))
@@ -555,7 +557,16 @@ function _resolve_merge_pip_packages(packages)
                 "direct references ('@ ...') and version specifiers both given for pip package '$name'",
             )
         end
-        push!(specs, PipPkgSpec(name, version = version, binary = binary, extras = extras))
+        push!(
+            specs,
+            PipPkgSpec(
+                name,
+                version = version,
+                binary = binary,
+                extras = extras,
+                editable = editable,
+            ),
+        )
     end
     sort!(specs, by = x -> x.name)
 end
@@ -669,9 +680,16 @@ function _resolve_pip_install(io, pip_specs, load_path, backend)
         elseif spec.binary == "no"
             push!(args, "--no-binary", spec.name)
         end
+        if spec.editable
+            # remove the @ from the beginning of the path.
+            url = replace(spec.version, r"@\s*" => "")
+            push!(args, "--editable", url)
+        end
     end
     for spec in pip_specs
-        push!(args, specstr(spec))
+        if !spec.editable
+            push!(args, specstr(spec))
+        end
     end
     vrb = _verbosity_flags()
     flags = vrb