+++ test/main.jl
@@ -164,10 +164,11 @@ end
 end
 
 @testitem "pip install/remove local python package" begin
-    @testset "file $file" for file in [
-        "example-python-package",
-        "example_python_package-1.0.0-py3-none-any.whl",
-        "example_python_package-1.0.0.tar.gz",
+    @testset "file $file $kwargs" for (file, kwargs) in [
+        ("example-python-package", NamedTuple()),
+        ("example_python_package-1.0.0-py3-none-any.whl", NamedTuple()),
+        ("example_python_package-1.0.0.tar.gz", NamedTuple()),
+        ("example-python-package", (editable = true,)),
     ]
         include("setup.jl")
         CondaPkg.add("python", version = "==3.10.2")
@@ -180,13 +181,46 @@ end
 
         # install package
         path = "./test/data/$file"
-        CondaPkg.add_pip("example-python-package", version = "@$path")
+        fullpath = abspath(dirname(CondaPkg.cur_deps_file()), path)
+        @assert ispath(fullpath)
+        editable = get(kwargs, :editable, false)
+        CondaPkg.add_pip("example-python-package", version = "@$path"; kwargs...)
         @test occursin("example-python-package", status())
-        @test occursin("(@$path)", status())
+        if isempty(kwargs)
+            @test occursin("(@$path)", status())
+        else
+            @test occursin("(@$path,", status())
+        end
+        @test occursin("editable", status()) == editable
         CondaPkg.withenv() do
             isnull || run(`python -c "import example_python_package"`)
         end
         @test occursin("v1.0.0", status()) == !isnull
+
+        # check editability
+        if editable && !isnull
+            @assert isdir(fullpath)
+            added_path = joinpath(fullpath, "src", "example_python_package", "added.py")
+            # check a particular submodule does not exist
+            rm(added_path; force = true)
+            CondaPkg.withenv() do
+                @test_throws Exception run(
+                    `python -c "from example_python_package.added import foo"`,
+                )
+            end
+            # now add it and check we can import it
+            write(added_path, "foo = 12")
+            CondaPkg.withenv() do
+                run(`python -c "from example_python_package.added import foo"`)
+            end
+            # remove it again
+            rm(added_path; force = true)
+            CondaPkg.withenv() do
+                @test_throws Exception run(
+                    `python -c "from example_python_package.added import foo"`,
+                )
+            end
+        end
 
         # remove package
         CondaPkg.rm_pip("example-python-package")