+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+* Add `editable` property to pip packages, specifying whether to install them in
+  editable mode.
+
 ## 0.2.31 (2025-08-28)
 * Bug fix: pip packages specified by file location are now correctly converted to "path"
   installs with Pixi.